+++ jest.config.js
@@ -94,6 +94,8 @@ module.exports = {
 
     // A preset that is used as a base for Jest's configuration
     // preset: undefined,
+    "preset": "@shelf/jest-mongodb",
+
 
     // Run tests from one or more projects
     // projects: undefined,
@@ -129,6 +131,7 @@ module.exports = {
 
     // A list of paths to modules that run some code to configure or set up the testing framework before each test
     // setupFilesAfterEnv: [],
+    "setupFilesAfterEnv": ["jest-extended"],
 
     // The number of seconds after which a test is considered as slow and reported as such in the results.
     // slowTestThreshold: 5,
@@ -174,6 +177,7 @@ module.exports = {
 
     // A map from regular expressions to paths to transformers
     // transform: undefined,
+    transform: {},
 
     // An array of regexp pattern strings that are matched against all source file paths, matched files will skip transformation
     // transformIgnorePatterns: [