+++ models/responses/schema.js
@@ -5,9 +5,23 @@
  */
 
 const mongoose = require('mongoose');
+const source = require('rfr');
+
+const { snowflakeValidator } = source('models/util');
 
 module.exports = new mongoose.Schema({
-    guild: String,
-    message: String,
-    botMode: String,
+    guild: {
+        type: String,
+        maxLength: 20,
+        required: true,
+        validate: snowflakeValidator,
+    },
+    message: {
+        type: String,
+        required: true,
+    },
+    botMode: {
+        type: String,
+        required: true,
+    },
 });