+++ models/role/schema.js
@@ -5,9 +5,23 @@
  */
 
 const mongoose = require('mongoose');
+const source = require('rfr');
+
+const { snowflakeValidator } = source('models/util');
 
 module.exports = new mongoose.Schema({
-    guild: String,
-    emoji_name: String,
-    role_name: String,
+    guild: {
+        type: String,
+        required: true,
+        maxLength: 20,
+        validate: snowflakeValidator,
+    },
+    emoji_name: {
+        type: String,
+        required: true,
+    },
+    role_name: {
+        type: String,
+        required: true,
+    },
 });