+++ routes/controllers/docsController.js
@@ -2,7 +2,8 @@ const source = require('rfr');
 const MarkdownIt = require('markdown-it');
 const fs = require('fs');
 
-const { setupCommands } = source('bot/scanner/commandHandler');
+const Bot = source('bot');
+
 const md = new MarkdownIt();
 
 const gettingStartedMD = fs.readFileSync('docs/GettingStarted.md', 'utf-8');
@@ -10,12 +11,14 @@ const configurationsMD = fs.readFileSync('docs/Configurations.md', 'utf-8');
 const roleReactionsMD = fs.readFileSync('docs/RoleReactions.md', 'utf-8');
 const spotifyPluginMD = fs.readFileSync('docs/SpotifyPlugin.md', 'utf-8');
 const gettingHelpMD = fs.readFileSync('docs/GettingHelp.md', 'utf-8');
+const webhooksMD = fs.readFileSync('docs/webhooks.md', 'utf-8');
 
 const gettingStarted = md.render(gettingStartedMD);
 const configurations = md.render(configurationsMD);
 const roleReactions = md.render(roleReactionsMD);
 const spotifyPlugin = md.render(spotifyPluginMD);
 const gettingHelp = md.render(gettingHelpMD);
+const webhooks = md.render(webhooksMD);
 
 function renderMarkdown(command) {
     return (command.docs) ? md.render(command.docs) : null;
@@ -23,15 +26,16 @@ function renderMarkdown(command) {
 
 const DocsController = {
     get(req, res) {
-        const client = {
-        };
-        setupCommands(client);
-        const commands = client.commands.map((command) => renderMarkdown(command)).filter(Boolean);
+        const commands = Bot.client.commands
+            .map((command) => renderMarkdown(command))
+            .filter(Boolean);
+
         res.render('docs', {
             gettingStarted,
             configurations,
             roleReactions,
             spotifyPlugin,
+            webhooks,
             commands,
             gettingHelp,
         });